load Win32API::File for _move_file_at_boot only when needed (#6)
Miniperl does not have Win32API::File available, first it is XS and not
miniperl compatible, second it is built very late in a parallel perl core
build. The eval will always fail with miniperl.

_move_file_at_boot is very rarely called, only when an XS .dll that is
loaded in a probably hung perl.exe process needs to be overwritten with a
a "make install" (if the blib dll is in use in a hung perl, the Win32 VC
or GCC linker will fatally error). This does not occur in a core perl
build. So to speed things up by not looking for a non-existant
Win32API::File on disk every time EU::I is loaded during the perl build,
and save memory/time by not loading Win32API::File for CPAN module building
only load it when it is first needed. Cache the result so require isn't
called over and over. Use a closure instead of a glob/global for speed.
Personally I'd remove the eval and let it visibly hard fail (your module
was not installed) since Win32API::File has been core distributed since
5.8.9, but I will be conservative here.