clean up POD structure
Mixing =begin sections inside and out of =over sections isn't really
valid.

Elevate the Functions section to a head1, bringing outside the
DESCRIPTION section. Change all of the function pod to be head2
sections, allowing them to appear in the table of contents. Also convert
the private and undocumented function pod to head2, which eliminates the
weirdness of mixing begin sections with over sections.

Remove the DISCOURAGED note from the heading of install_default, to
improve linking.